requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Django==4.2.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
multidict==6.0.4
Pillow==9.5.0
pyasn1==0.5.0
pycparser==2.21
pypng==0.20220715.0
python-dotenv==1.0.0
python-telegram-bot==20.3
qrcode==7.4.2
requests==2.31.0
rsa==4.9
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.2
yarl==1.9.2
